+++ crates/zizmor/src/registry.rs
@@ -85,6 +85,7 @@ impl AuditRegistry {
         register_audit!(audit::unpinned_tools::UnpinnedTools);
         register_audit!(audit::adhoc_packages::AdhocPackages);
         register_audit!(audit::insecure_url_scheme::InsecureURLScheme);
+        register_audit!(audit::self_repository::SelfRepository);
 
         Ok(registry)
     }